Defining a new type called author, which is a struct. A struct is a composite data type that can hold multiple values of different types. In this case, the author struct has two fields: first and last, which are both strings. Strings are sequences of characters that can represent text.
The next block of code defines a method called fullName for the author type. A method is a function that can be called on a value of a specific type, using the dot notation. For example, if a is an author, then a.fullName() calls the fullName method on a. The method takes an argument called a, which is an author value, and returns a string that is the concatenation of the first and last fields of a, separated by a space. The syntax (a author) before the method name indicates that this method belongs to the author type, and that a is the receiver of the method. A receiver is similar to this or self in other languages, but it can have any name and does not need to be a pointer.

The next block of code defines the main function, which is the entry point of the program. The main function does not take any arguments or return any values. It simply executes the statements inside its body.

The first statement inside the main function creates a new variable called a, and assigns it an author value. The syntax {first: "Marcus", last: "Aurelius"} is called a struct literal, which creates a new struct value with the specified field values. This is similar to object literals in JavaScript or dictionaries in Python

The second statement inside the main function calls the fmt.Println function, which prints its arguments to the standard output, followed by a newline. In this case, the argument is a.fullName(), which calls the fullName method on a, and returns the string "Marcus Aurelius".